Improve the db setting extract constant name
