+++ lib/parliamentarians/extract/db_extract.rb
@@ -1,26 +1,18 @@
-require 'mongo'
+require '../../mongo_db'
+require '../../settings'
 
-# This class is used to connect with mongo
+# This class is used to connect with mongo extract db
 #
 # Example of use:
-#   @db = Parliamentarians::DbExtract.new(EXTRACT_MONGO_SETTINGS[env])
+#   @db = Parliamentarians::DbExtract.new(EXTRACT_MONGO_DB_SETTINGS[env])
 #   @db.save('your-data-here')
 
 module Parliamentarians
-  class DbExtract
-    def initialize(settings)
-      host = settings['host'] + ':' + settings['port']
+  class DbExtract < DB
+    def initialize
+      host = EXTRACT_MONGO_DB_SETTINGS['host'] + ':' + EXTRACT_MONGO_DB_SETTINGS['port']
       @@client = Mongo::Client.new([host], :database => settings['name'])
       @collection = @@client[:parliamentarians]
-    end
-
-    def save(doc)
-      result = @collection.insert_one(doc)
-      result.n
-    end
-
-    def get(search)
-      @collection.find(search).first
     end
   end
 end
